requirements.txt
# Web framework & server
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# HTTP client
httpx==0.27.0

# Scheduler
APScheduler==3.10.4

# Image processing (optional but recommended)
Pillow==10.3.0

# Video processing (optional – requires ffmpeg binary)
moviepy==1.0.3

# MinIO client (object storage)
minio==7.2.7

# Redis client (optional caching)
redis==5.0.4

# Async utilities
anyio==4.3.0
